FixingVerticallity
I made the camera follow the player vertically, and I fixed a problem with player jumping where when landing, you would be in the ground so you would jump for a longer time getting a higher vertical velocity. Also, I fixed the test scene to be centered at 0 for convienent level layout in the files.