RemovingWallJump
I made it so the contact points need to be within the x width of the player collider so you can't jump off walls. I also reduced the force of movement when off the ground so it's harder to turn around mid jump, and so you can't cling to a wall completely by pushing into it. However, now onGround isn't set to true if the player lands and isn't moving horizontally, and because of that, along with the reduced force when onGround is false, the player movement is overly slow at the start. Also, if you land without moving you can't jump until you move.