FixingOnGroundBug
Now you can move normally, and jump from a stand still. You still can't cling to the wall, and you can't jump up a wall. However, if you're on the ledge of the wall you can jump up off that. I don't know that that's necessarily a problem so I'm just going to leave it as it is, but if you want to fix that, have fun trying to fix it without breaking anything else.