FixingOnGroundBug
onGround is now set correctly. You can move normally, and jump from a standstill. Also, you still can't cling to walls, or jump up a wall. However, you can jump if you're on the very top edge of the wall. I don't know if that's a problem, so I'm not going to mess with it any more. If you want to fix it, have fun trying to remove that feature, without breaking any other feature.