FixingPlayerStopMove
When you walk off one instance of the ground to the next, you leave a ground collider setting onGround to false, which is then not updated for some reason. I don't think this fixes it quite right, but I'm not sure what's going on and this was the best I could come up with.