ChangingHowThingsAreSpawnedAndDespawned
I made it so objects are spawned when within 2 rooms of the player and de spawned when further away. Also, I made it so that projectile shooters have a range that they pass to their bullets. Finally I made a wall prefab to keep the player in between the start and end. Fun challenge, try to fly over the wall on the right using the bullets hitting the back of the shield.