ChangingLevelLoading+LevelEditor
I changed how we load rooms. Now all of them are instantiated at the start, and activated or deactivated as you progress through the level to make it smoother as you move through the level. Also, I made a very simple level editor. In the level editor scene there is a box displaying the width of the room. For the most part all objects should be placed within the box. Make sure that when you place the prefab into the rooms folder in resources you give it the same name in both the roomData field in the inspector and the object itself. Once a room is placed in the folder, I've made a new menu tool drop down called ESTools. Click the SaveRooms button and it'll update the roomData.csv file with the rooms we have.