FixingBackdrop+ImprovingRoomCreator
I made the backdrop class use arrays instead of distinct objects, it now looks much neater, and allows for any number of backdrop objects. I also added a button to the inspector for level creation which makes it so you don't need to manually put the room in the rooms resources folder, and automatically uses the correct name so you don't need to put the same name in two places.