MadeReflectionPossiblyBetter
Reflection now no longer takes the direction of your velocity to bounce you. Instead you always go backwards from the way your shield is pointing. Even if that's not really realistic, I think that makes the game play better. It does make it impossible to fly forever on projectiles, which is why there is a screen shot here. The last time the bird makes it to the top of the tower.
I also made it so projectiles reflect such that their angle of incidence is the same as the exit angle when moving towards the shield. When they move away from the shield, they instead move in the direction the shield is pointing.