PlayerData
I added the PlayerData.csv file and the method to read it. I also modified the dash speed to have an x and y component which we can adjust independently. If you want the player to use a player data, use the character field in the player. 0 is the current test player, 1 is a different one I made for a test bird that can't dash or crouch, but can jump 5 times.