StuffChanged
I added an extra room to the third world.   Also, I made the room tester have a horizontal displacement, so you can spawn anywhere in the level. Finally, I made it so the bird is better. For some reason I had it so when you jump, it didn't reset the vertical velocity when moving down.