CameraReflectionProjectileRework
Made it so reflecting off a projectile no longer used the player speed, to make for more consistant levels. Then adjusted projectile masses to fit the new code. Changed camera to fit within the room min and max X to make the rooms seperate, like in Celeste, or hollow knight, so the player knows where they're going to respawn if they fall out of the world.