ChangedDeath
I know we have a great lose screen. However, with how I'm moving the game i think it makes more sense for the game to have instant death when you fall in a pit, but death just restarts the room, rather than starting the game over from the start. So we'll instead record number of deaths, for people who want to complete it with as few as possible, in addition to time.