AddedNewProjectileTypes
Now there's shieldbreaker projectiles and vaccum projectiles. ShieldBreaker objects aren't reflected by the shiled but still do damage to the player, so they have to be avoided. Vaccum projectiles will be reflected the opposite direction as the shield is pointing, and instead of bouncing the player away from it, it pulls them towards it.